% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{a5_cell_area}
\alias{a5_cell_area}
\title{Cell area at a given resolution}
\usage{
a5_cell_area(resolution)
}
\arguments{
\item{resolution}{Integer vector of resolutions (0--30).}
}
\value{
A numeric vector of areas in square metres.
}
\description{
Returns the area of a single cell in square metres at the given
resolution(s). Because A5 is an equal-area DGGS, all cells at the same
resolution have identical area.
}
\examples{
a5_cell_area(0:5)
}
