% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{a5_cell_area_rs}
\alias{a5_cell_area_rs}
\title{Get the area (in square metres) of cells at a given resolution.}
\usage{
a5_cell_area_rs(resolution)
}
\arguments{
\item{resolution}{Integer vector of resolutions (0-30).}
}
\value{
Numeric vector of areas in square metres.
}
\description{
Get the area (in square metres) of cells at a given resolution.
}
