% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{a5_cell_to_boundary_rs}
\alias{a5_cell_to_boundary_rs}
\title{Get boundary polygon vertices for A5 cells.}
\usage{
a5_cell_to_boundary_rs(cell, closed_ring, segments)
}
\arguments{
\item{cell}{Character vector of hex-encoded cell IDs.}

\item{closed_ring}{Logical: should the polygon ring be closed?}

\item{segments}{Integer: number of interpolation segments per edge.}
}
\value{
A list of lists, each with \code{lon} and \code{lat} numeric vectors.
}
\description{
Get boundary polygon vertices for A5 cells.
}
