% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{a5_cell_to_children_rs}
\alias{a5_cell_to_children_rs}
\title{Get child cells.}
\usage{
a5_cell_to_children_rs(cell, child_resolution)
}
\arguments{
\item{cell}{A single hex-encoded cell ID.}

\item{child_resolution}{Integer: target child resolution. NULL for
immediate children.}
}
\value{
Character vector of hex-encoded child cell IDs.
}
\description{
Get child cells.
}
