% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{a5_cell_to_parent_rs}
\alias{a5_cell_to_parent_rs}
\title{Navigate to parent cell(s).}
\usage{
a5_cell_to_parent_rs(cell, parent_resolution)
}
\arguments{
\item{cell}{Character vector of hex-encoded cell IDs.}

\item{parent_resolution}{Integer: target parent resolution. NULL for
immediate parent.}
}
\value{
Character vector of hex-encoded parent cell IDs.
}
\description{
Navigate to parent cell(s).
}
