% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{a5_compact_rs}
\alias{a5_compact_rs}
\title{Compact a set of A5 cell IDs.}
\usage{
a5_compact_rs(cells)
}
\arguments{
\item{cells}{Character vector of hex-encoded cell IDs.}
}
\value{
Character vector of compacted hex-encoded cell IDs.
}
\description{
Merges sibling groups into their common parent.
}
