% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a5r.R
\name{a5_from_hex}
\alias{a5_from_hex}
\title{Convert hexadecimal string to A5 cell ID}
\usage{
a5_from_hex(hex)
}
\arguments{
\item{hex}{Hexadecimal string representation of the cell ID.}
}
\value{
A5 cell ID as an \code{integer64} value.
}
\description{
Convert hexadecimal string to A5 cell ID
}
\examples{
a5_from_hex("63611a8000000000")
}
