% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a5r.R
\name{a5_from_lonlat}
\alias{a5_from_lonlat}
\title{Convert longitude/latitude to A5 cell ID}
\usage{
a5_from_lonlat(lon, lat, resolution)
}
\arguments{
\item{lon}{Longitude in decimal degrees.}

\item{lat}{Latitude in decimal degrees.}

\item{resolution}{Resolution level of the grid (0-30).}
}
\value{
A5 cell ID as an \code{integer64} value.
}
\description{
Convert longitude/latitude to A5 cell ID
}
\examples{
cell <- a5_from_lonlat(-0.16, 51.51, 10)
}
