% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{a5_get_num_cells_rs}
\alias{a5_get_num_cells_rs}
\title{Get total number of cells at a given resolution.}
\usage{
a5_get_num_cells_rs(resolution)
}
\arguments{
\item{resolution}{Integer scalar (0-30).}
}
\value{
Numeric scalar (as double, since R has no u64).
}
\description{
Get total number of cells at a given resolution.
}
