% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{a5_get_res0_cells}
\alias{a5_get_res0_cells}
\title{Get all resolution-0 root cells}
\usage{
a5_get_res0_cells()
}
\value{
An \link{a5_cell} vector of length 12.
}
\description{
Returns the 12 root cells corresponding to the 12 faces of the
dodecahedron.
}
\examples{
a5_get_res0_cells()
}
