% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{a5_get_res0_cells_rs}
\alias{a5_get_res0_cells_rs}
\title{Get all 12 resolution-0 root cells.}
\usage{
a5_get_res0_cells_rs()
}
\value{
Character vector of 12 hex-encoded cell IDs.
}
\description{
Get all 12 resolution-0 root cells.
}
