% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{a5_get_resolution}
\alias{a5_get_resolution}
\title{Get the resolution of A5 cell indices}
\usage{
a5_get_resolution(cell)
}
\arguments{
\item{cell}{An \link{a5_cell} vector.}
}
\value{
An integer vector of resolutions.
}
\description{
Extracts the resolution level (0--30) encoded in each cell index.
}
\examples{
cell <- a5_lonlat_to_cell(-3.19, 55.95, resolution = 10)
a5_get_resolution(cell)
}
\seealso{
\code{\link[=a5_cell_to_parent]{a5_cell_to_parent()}}, \code{\link[=a5_cell_to_children]{a5_cell_to_children()}}
}
