% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{a5_get_resolution_rs}
\alias{a5_get_resolution_rs}
\title{Get the resolution of A5 cell indices.}
\usage{
a5_get_resolution_rs(cell)
}
\arguments{
\item{cell}{Character vector of hex-encoded cell IDs.}
}
\value{
Integer vector of resolutions.
}
\description{
Get the resolution of A5 cell indices.
}
