% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{a5_hex_to_cell}
\alias{a5_hex_to_cell}
\title{Convert hex string to A5 cell}
\usage{
a5_hex_to_cell(hex)
}
\arguments{
\item{hex}{Character vector of hex-encoded cell IDs.}
}
\value{
An \link{a5_cell} vector.
}
\description{
Convenience wrapper to construct an \link{a5_cell} from a hex string.
}
\examples{
a5_hex_to_cell("0800000000000006")
}
