% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{a5_is_cell}
\alias{a5_is_cell}
\title{Test if values are valid A5 cell indices}
\usage{
a5_is_cell(x)
}
\arguments{
\item{x}{An \link{a5_cell} vector or character vector of hex strings.}
}
\value{
A logical vector.
}
\description{
Checks whether each element is a syntactically valid hex-encoded A5 cell
ID.
}
\examples{
a5_is_cell(c("0800000000000006", "not_a_cell", NA))
}
