% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{a5_is_valid_cell_rs}
\alias{a5_is_valid_cell_rs}
\title{Validate hex cell IDs.}
\usage{
a5_is_valid_cell_rs(cell)
}
\arguments{
\item{cell}{Character vector of hex strings to validate.}
}
\value{
Logical vector indicating validity.
}
\description{
Validate hex cell IDs.
}
