% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{a5_lonlat_to_cell}
\alias{a5_lonlat_to_cell}
\title{Convert coordinates to A5 cell indices}
\usage{
a5_lonlat_to_cell(lon, lat, resolution)
}
\arguments{
\item{lon}{Numeric vector of longitudes in degrees.}

\item{lat}{Numeric vector of latitudes in degrees.}

\item{resolution}{Integer scalar or vector of resolutions (0--30).}
}
\value{
An \link{a5_cell} vector of cell indices.
}
\description{
Maps longitude/latitude coordinates to A5 cell indices at the specified
resolution.
}
\examples{
a5_lonlat_to_cell(-3.19, 55.95, resolution = 5)
}
\seealso{
\code{\link[=a5_cell_to_lonlat]{a5_cell_to_lonlat()}} for the inverse operation.
}
