% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a5r.R
\name{a5_resolution}
\alias{a5_resolution}
\title{Get the resolution of an A5 cell}
\usage{
a5_resolution(cell)
}
\arguments{
\item{cell}{A5 cell ID as an \code{integer64} value.}
}
\value{
Resolution level (0-30) of the cell.
}
\description{
Get the resolution of an A5 cell
}
\examples{
cell <- a5_from_lonlat(-0.16, 51.51, 10)
a5_resolution(cell)
}
