% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a5r.R
\name{a5_to_hex}
\alias{a5_to_hex}
\title{Convert A5 cell ID to hexadecimal string}
\usage{
a5_to_hex(cell)
}
\arguments{
\item{cell}{A5 cell ID as an \code{integer64} value.}
}
\value{
Hexadecimal string representation of the cell ID.
}
\description{
Convert A5 cell ID to hexadecimal string
}
\examples{
cell <- a5_from_lonlat(-0.16, 51.51, 10)
a5_to_hex(cell)
}
