% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a5r.R
\name{a5_to_lonlat}
\alias{a5_to_lonlat}
\title{Convert A5 cell ID to longitude/latitude}
\usage{
a5_to_lonlat(cell)
}
\arguments{
\item{cell}{A5 cell ID as an \code{integer64} value.}
}
\value{
A list with \code{lon} and \code{lat} components.
}
\description{
Convert A5 cell ID to longitude/latitude
}
\examples{
cell <- a5_from_lonlat(-0.16, 51.51, 10)
a5_to_lonlat(cell)
}
