% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{a5_uncompact_rs}
\alias{a5_uncompact_rs}
\title{Uncompact a set of A5 cell IDs to a target resolution.}
\usage{
a5_uncompact_rs(cells, target_resolution)
}
\arguments{
\item{cells}{Character vector of hex-encoded cell IDs.}

\item{target_resolution}{Integer: the resolution to expand to.}
}
\value{
Character vector of uncompacted hex-encoded cell IDs.
}
\description{
Uncompact a set of A5 cell IDs to a target resolution.
}
