% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{blob_to_hex_rs}
\alias{blob_to_hex_rs}
\title{Convert a list of raw(8) blobs to hex strings (zero-padded to 16 chars).}
\usage{
blob_to_hex_rs(cells)
}
\description{
Convert a list of raw(8) blobs to hex strings (zero-padded to 16 chars).
}
