% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{cell_to_children}
\alias{cell_to_children}
\title{Get the children cells at a finer resolution}
\usage{
cell_to_children(cell_id, child_resolution)
}
\arguments{
\item{cell_id}{ID of the cell.}

\item{child_resolution}{target resolution (must be greater than current), or NULL for immediate children.}
}
\value{
vector of child cell IDs.
}
\description{
Get the children cells at a finer resolution
}
