% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{cell_to_lon_lat}
\alias{cell_to_lon_lat}
\title{Convert A5 cell ID to latitude/longitude}
\usage{
cell_to_lon_lat(cell_id)
}
\arguments{
\item{cell_id}{ID of the cell.}
}
\description{
Convert A5 cell ID to latitude/longitude
}
