% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{cell_to_parent}
\alias{cell_to_parent}
\title{Get the parent cell at a coarser resolution}
\usage{
cell_to_parent(cell_id, parent_resolution)
}
\arguments{
\item{cell_id}{ID of the cell.}

\item{parent_resolution}{target resolution (must be less than current), or NULL for immediate parent.}
}
\value{
parent cell ID.
}
\description{
Get the parent cell at a coarser resolution
}
