% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compact}
\alias{compact}
\title{Compact a set of cells by merging complete sibling groups into parent cells}
\usage{
compact(cells)
}
\arguments{
\item{cells}{vector of cell IDs to compact.}
}
\value{
vector of compacted cell IDs (typically smaller than input).
}
\description{
Compact a set of cells by merging complete sibling groups into parent cells
}
