% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_cell_area}
\alias{get_cell_area}
\title{Get the area of a cell at a given resolution}
\usage{
get_cell_area(resolution)
}
\arguments{
\item{resolution}{resolution level (0-30).}
}
\value{
area in square meters.
}
\description{
Get the area of a cell at a given resolution
}
