% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_num_cells}
\alias{get_num_cells}
\title{Get the number of cells at a given resolution}
\usage{
get_num_cells(resolution)
}
\arguments{
\item{resolution}{resolution level (0-30).}
}
\value{
number of cells at the specified resolution.
}
\description{
Get the number of cells at a given resolution
}
