% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_res0_cells}
\alias{get_res0_cells}
\title{Get all resolution 0 (base) cells}
\usage{
get_res0_cells()
}
\value{
vector of all 12 base cell IDs.
}
\description{
Get all resolution 0 (base) cells
}
