% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_resolution}
\alias{get_resolution}
\title{Get the resolution of an A5 cell}
\usage{
get_resolution(cell_id)
}
\arguments{
\item{cell_id}{ID of the cell.}
}
\value{
resolution level (0-30) of the cell.
}
\description{
Get the resolution of an A5 cell
}
