% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{hex_to_u64}
\alias{hex_to_u64}
\title{Convert hexadecimal string to A5 cell ID}
\usage{
hex_to_u64(hex)
}
\arguments{
\item{hex}{hexadecimal string representation of the cell ID.}
}
\value{
A5 cell ID as a numeric value.
}
\description{
Convert hexadecimal string to A5 cell ID
}
