% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{lon_lat_to_cell}
\alias{lon_lat_to_cell}
\title{Convert latitude/longitude to A5 cell ID}
\usage{
lon_lat_to_cell(lon, lat, resolution)
}
\arguments{
\item{lon}{longitude in decimal degrees.}

\item{lat}{latitude in decimal degrees.}

\item{resolution}{resolution level of the grid}
}
\description{
Convert latitude/longitude to A5 cell ID
}
