% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{u64_to_hex}
\alias{u64_to_hex}
\title{Convert A5 cell ID to hexadecimal string}
\usage{
u64_to_hex(cell_id)
}
\arguments{
\item{cell_id}{ID of the cell.}
}
\value{
hexadecimal string representation of the cell ID.
}
\description{
Convert A5 cell ID to hexadecimal string
}
