% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{uncompact}
\alias{uncompact}
\title{Uncompact (expand) cells to a target resolution}
\usage{
uncompact(cells, target_resolution)
}
\arguments{
\item{cells}{vector of cell IDs to expand.}

\item{target_resolution}{the target resolution for all output cells.}
}
\value{
vector of cell IDs all at the target resolution.
}
\description{
Uncompact (expand) cells to a target resolution
}
