% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell.R
\name{a5_cell}
\alias{a5_cell}
\alias{is_a5_cell}
\alias{as_a5_cell}
\alias{vec_ptype_abbr.a5_cell}
\alias{vec_ptype_full.a5_cell}
\alias{pillar_shaft.a5_cell}
\title{A5 Cell Index Vector}
\usage{
a5_cell(x = character())

is_a5_cell(x)

as_a5_cell(x)

vec_ptype_abbr.a5_cell(x, ...)

vec_ptype_full.a5_cell(x, ...)

pillar_shaft.a5_cell(x, ...)
}
\arguments{
\item{x}{A character vector of hex-encoded A5 cell IDs, or an object
coercible to one.}
}
\value{
An \code{a5_cell} vector.
}
\description{
A vctrs-based vector type representing A5 cell indices. Cell IDs are
stored as hex-encoded character strings internally. The \code{a5_cell} type
provides type safety, pretty-printing, and integration with the vctrs
framework.
}
\examples{
cells <- a5_cell(c("0800000000000006", "0800000000000016"))
cells
}
