% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{a5_cell_to_children}
\alias{a5_cell_to_children}
\title{Get child cells}
\usage{
a5_cell_to_children(cell, resolution = NULL)
}
\arguments{
\item{cell}{A single \link{a5_cell} value.}

\item{resolution}{Integer scalar target child resolution, or \code{NULL} for
immediate children.}
}
\value{
An \link{a5_cell} vector of child cells.
}
\description{
Returns the child cells of a single cell. By default returns the 4
immediate children (one resolution finer). Optionally target a specific
finer resolution.
}
\examples{
cell <- a5_lonlat_to_cell(-3.19, 55.95, resolution = 5)
a5_cell_to_children(cell)
}
