% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{a5_cell_to_lonlat}
\alias{a5_cell_to_lonlat}
\title{Convert A5 cell indices to coordinates}
\usage{
a5_cell_to_lonlat(cell)
}
\arguments{
\item{cell}{An \link{a5_cell} vector (or character coercible to one).}
}
\value{
A \code{\link[wk:xy]{wk::xy()}} vector of (longitude, latitude) points.
}
\description{
Returns the centre-point longitude and latitude of each cell.
}
\examples{
cell <- a5_lonlat_to_cell(-3.19, 55.95, resolution = 5)
a5_cell_to_lonlat(cell)
}
