% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a5R-package.R
\docType{package}
\name{a5R-package}
\alias{a5R}
\alias{a5R-package}
\title{a5R: A5 Discrete Global Grid System}
\description{
R bindings for the A5 Discrete Global Grid System (DGGS). A5 partitions the Earth's surface into pentagonal cells across 31 resolution levels using an equal-area projection onto a dodecahedron. Provides functions for indexing coordinates to cells, traversing the cell hierarchy, computing cell boundaries, and compacting/uncompacting cell sets. Powered by the 'a5' Rust crate via 'extendr'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/belian-earth/a5R}
  \item \url{https://belian-earth.github.io/a5R/}
  \item Report bugs at \url{https://github.com/belian-earth/a5R/issues}
}

}
\author{
\strong{Maintainer}: Hugh Graham \email{hugh.a.graham@gmail.com}

}
\keyword{internal}
