% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wk.R
\name{wk_methods}
\alias{wk_methods}
\alias{wk_handle.a5_cell}
\alias{wk_crs.a5_cell}
\title{wk methods for a5_cell}
\usage{
\method{wk_handle}{a5_cell}(handleable, handler, ...)

\method{wk_crs}{a5_cell}(x)
}
\arguments{
\item{handleable, x}{An \link{a5_cell} vector.}

\item{handler}{A \link[wk:wk_handle]{wk handler}.}

\item{...}{Passed to underlying methods.}
}
\value{
\itemize{
\item \code{wk_handle()}: the result of the handler.
\item \code{wk_crs()}: a \link[wk:wk_crs]{wk::wk_crs} object (WGS 84 lon/lat).
}
}
\description{
Integration with the \link[wk:wk-package]{wk} geometry framework. Allows
\code{a5_cell} vectors to be handled as geometry (via their boundary polygons)
and to report their CRS.
}
